DRY on unittests
